Add C-k to tmux to clear scrollback buffer
The shortcut `C-l` just clears the screen but retains the scrollback
buffer. In the case of dealing with lots of output from a command like
strace, it's useful to clear the scrollback buffer so that you know any
searches performed in the terminal session are not part of a previous
command's output. This commit implements clearing the screen and the
current sessions scrollback buffer.