Add FZF ENV variables to use silversearcher
Sets FZF_DEFAULT_COMMAND and FZF_CTRL_T_COMMAND and ignores `.git` and
`node_module` folders by default. `ag` also honours `.gitignore` files
but you need to be in the root folder that it exists for it to work. By
ignoring `.git` and `node_modules` by default, this will enable it to
work properly when I utilise it from within `vim`.