Add `sshsplit` to sync commands accross multiple ssh hosts
sshsplit - A bash script for creating tmux windows, each with multiple
panes. Each pane SSH's into a different server based on a pattern in the
SSH config file.

```
 Usage: ./sshsplit <ssh host pattern> [range] [max splits]
```

This script takes a pattern and an optional range as command line
arguments. The pattern is used to match hostnames in the SSH config
file. If a range is provided, the script only considers hostnames with
numbers within that range. The pattern and range should match the format
used for hostnames in the SSH config file.

The script creates a new tmux window for every 5 matched hosts (this
number can be changed by modifying the `max_splits` variable). In each
window, it creates a pane for each host and initiates an SSH connection.

Each window is named according to the pattern and the range of host
numbers in that window, which makes it easy to identify the servers
connected in each window. The script also synchronizes the panes in each
window, meaning that typing in one pane sends the input to all panes in
the same window. This is useful for running the same commands on
multiple servers simultaneously.

```
 Host server1
   HostName 192.168.1.101
   Port 22
   User your_username
   IdentityFile ~/.ssh/id_rsa
```
For example, if you have servers named "server1", "server2", ...,
"server20" in your SSH config file, you can use the command "./sshsplit
server 1-20" to create 4 tmux windows, each with 5 panes connected to 5
different servers.